requirements.txt
fastapi
uvicorn
sqlalchemy
jinja2
python-multipart
openai
pydantic
python-dotenv
aiofiles
httpx
